requirements.txt
python-amazon-sp-api==2.1.8
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
apscheduler==3.10.4
rich==13.7.1
lxml>=5.3.0
